Extract task management to a separate module (no changes, just a move)
